mini_ipc: fix bugs

+++ src/wii/mini_ipc.c
@@ -37,8 +37,8 @@ struct mini_state {
 	struct infohdr *infohdr;                      /* MINI infohdr */
 	u16 out_head;                                 /* Out head */
 	u16 in_tail;                                  /* In tail */
-	int in_size;                                  /* In queue size */
-	int out_size;                                 /* Out queue size */
+	u32 in_size;                                  /* In queue size */
+	u32 out_size;                                 /* Out queue size */
 	volatile struct ipc_request_mini *in_queue;   /* In queue pointer */
 	volatile struct ipc_request_mini *out_queue;  /* Out queue pointer */
 	u32 cur_tag;                                  /* Current request number ("tag") */
@@ -90,8 +90,8 @@ void MINI_Init(void) {
 
 	/* read in_tail and out_head from IPC hardware registers */
 	ppcmsg = HW_IPC_PPCMSG;
-	state.in_tail = ppcmsg & 0xffff;
-	state.out_head = ppcmsg >> 16;
+	state.in_tail = (u16)(ppcmsg & 0xffff);
+	state.out_head = (u16)(ppcmsg >> 16);
 
 	state.cur_tag = 1;
